[level 0] Title: 배열의 평균값, Time: 0.03 ms, Memory: 83.7 MB -BaekjoonHub

+++ 프로그래머스/0/120817. 배열의 평균값/배열의 평균값.java
@@ -5,7 +5,7 @@ class Solution {
         for(i=0; i < numbers.length; i++){
             sum += numbers[i];
         }
-        double average = (double) sum / numbers.length;
+        double average = sum / numbers.length;
         return average;
     }
 }
